# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Management API",
    "version": "1.0.0",
    "description": "API for managing students and courses with MongoDB"
  },
  "servers": [
    {
      "url": "https://student-management-a28c.onrender.com",
      "description": "Production API server"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID from OAuth provider"
          },
          "displayName": {
            "type": "string",
            "description": "User's display name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "provider": {
            "type": "string",
            "description": "OAuth provider (e.g., 'google')"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Authentication status message"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    }
  },
  "paths": {
...
    "/auth/google": {
      "get": {
        "summary": "Initiate Google OAuth login",
        "description": "Redirects to Google OAuth for authentication",
        "responses": {
          "302": {
            "description": "Redirect to Google OAuth"
          },
          "500": {
            "description": "OAuth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "description": "Handles the callback from Google OAuth after authentication",
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "OAuth not configured or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}